support image gradients as input
HOG is a function of image gradients. As such, the image itself is not
necessary for computing HOG features. Therefore, allow the user to pass
image gradients directly as a tuple. This, in turn, enables the use of
gradient estimation methods other than central (or discrete point)
differences.

As part of this rework, OpenCV central differences gradients were
replaced by Eigen Tensor based discrete point differences computation.
This removes the restrictions that stem from OpenCV's cv::Mat, such as
missing support for long double, among others. At the same time, the
dependency on OpenCV is loosened allowing to remove it entirely in the
future.